Reorganization of file and first preliminar version of Storybook for the component 'Select'
