--- connectmate/src/components/Layouts/Select/Select.stories.tsx
@@ -1,1 +1,0 @@
-export {}
