+++ connectmate/src/components/Layouts/Sidebar.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import Select from './Select/Select';
+import Select from '../Select/Select';
 
 const Sidebar: React.FC = () => {
   return (
