+++ connectmate/src/pages/HomePage.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 import { ROUTES } from '../ROUTES';
-import Select from '../components/Layouts/Select/Select';
+import Select from '../components/Select/Select';
 //import { MetricCard } from '../components/MetricCard';
 
 const HomePage: React.FC = () => {
